Creates EUPP Config used as Data Descriptor

Generates an object of class eupp_config Containing all the required settings for downloadin data set inventories (index files) and the data itself.

eupp_config(
  type = c("reforecast", "forecast", "analysis"),
  kind = c("ctr", "ens", "hr"),
  level = c("surf", "pressure", "efi"),
  date,
  parameter = NULL,
  steps = NULL,
  members = NULL,
  area = NULL,
  cache = NULL,
  version = 0L
)

Arguments

type

character of length 1 (see Details).

kind

character of length 1 or NULL (see Details).

level

character of length 1 (see Details).

date

an object which can be converted to POSIXct. Character string in ISO representation or an object of class Date or POSIXt.

parameter

NULL (default) or character. When NULL all available parameters will be processed. Can be used to retrieve/process subsets.

steps

NULL (default) or integer. If set to NULL all available forecast members will be processed. An integer sequence can be provided to only process specific forecast members.

area

NULL (default) or an object of class bbox. Used for spatial subsetting.

cache

NULL or character of length 1 pointing to an existing directory. Is used for data caching (caching grib index information/inventories) which can be handy to save some time.

version

integer lengt 1; version of the data set. Defaults to 0L. Typically not changed by an end-user.

Value

Returns an object of class eupp_config.

Details

Input argument type:

"reforecast": Reforecasts (or hindcasts).

"forecast": Forecast data.

"analysis": Analysis data; based on ECMWF ERA5 used as ground truth.

Input argument kind. Will be ignored if level = "efi" (has no kind).

"ctr": Control run.

"ens": Ensemble members (allows for optional argument members).

"hr": High-resolution forecast.

Input argument level:

"surf": Surface data.

"pressure": Pressure level data.

"efi": Extreme forecast index data.

Note that reforecasts always only become available on Mondays and Thursdays. If type = "reforecast" and the date does not point to Mon/Thu the function will throw an error.

The forecast ensemble (type = "forecast", kind = "ctr") consists of 50 ensemble members (1-50), the reforecasts or hindcasts (type = "forecast", kind = "ctr") only consists of 20 ensemble runs (1-20).

Author

Reto Stauffer

Examples

conf_A <- eupp_config(type = "reforecast", kind = "ctr", level = "surf",
                      date = "2017-01-02", parameter = "2t")

conf_B <- eupp_config(type = "reforecast", kind = "ctr", level = "surf",
                      date = "2017-01-02", parameter = "2t",
                      steps = c(6, 12))