Downloading EUPP Datasets

Main function downloading data.

Requires ncdf4 and the NetCDF Library to be installed.

eupp_download_gridded(
  x,
  output_file,
  output_format = c("grib", "nc"),
  verbose = FALSE,
  overwrite = FALSE
)

eupp_get_gridded(x, verbose = FALSE)

eupp_get_inventory(x, verbose = FALSE)

Arguments

x

object of class eupp_config.

output_file

character length 1, name of the output file.

output_format

character length 1 ("grib" or "nc"), defaults to "grib" (see details).

verbose

logical, sets verbosity level. Defaults to FALSE.

overwrite

logical length 1, defaults to FALSE. If set to TRUE the output_file will be overwritten if needed. If FALSE and output_file exists an error will be raised.

Value

Object of class stars; see Details.

Details

The function allows to store data sets in either GRIB version 1 or NetCDF (classic 64bit; v3). The original data set is provided as GRIB, the conversion to NetCDF is done locally usning ECMWFs ecCodes tools which must be installed when using NetCDF.

This function interfaces eupp_download_gridded to download the data and converts the original data set (GRIB version 1) to NetCDF. Thus NetCDF support and ecCodes tools are required.

If x (eupp_config) contains a bbox definition on area the data is subsetted directly.

See also

Author

Reto Stauffer

Reto Stauffer